Adds signup POST and error handling
Previously, the signup page did not respond to POST requests.
Now, it validates the posted form data, and either redirects
to the user page, or rerenders the signup form with error messages.

* Fetch the expected form parameters
* Use detect_errors to validate the data
* If detect_errors returns valid, redirect to the user page
* If not, pass the error object to the signup template
  and rerender

The user route handler retrieves the username parameter
from the url, and displays it. This is a stand-in for session
cookies that will be implemented later.